+++ java/us/thinkplank/grimcraft/item/ItemPhoenixEgg.java
@@ -13,8 +13,6 @@ import net.minecraft.util.EnumParticleTypes;
 import net.minecraft.util.SoundCategory;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
 import us.thinkplank.grimcraft.mobs.EntityPhoenix;
 
 public class ItemPhoenixEgg extends Item {
@@ -35,7 +33,7 @@ public class ItemPhoenixEgg extends Item {
 					worldIn.setBlockState(pos, Blocks.AIR.getDefaultState());
 					worldIn.playSound((EntityPlayer)null, pos.getX(), pos.getY(), pos.getZ(), SoundEvents.ENTITY_GENERIC_EXPLODE, SoundCategory.BLOCKS, 4.0F, (1.0F + (worldIn.rand.nextFloat() - worldIn.rand.nextFloat()) * 0.2F) * 0.7F);
 					EntityPhoenix phoenix = new EntityPhoenix(worldIn);
-					phoenix.setLocationAndAngles((double)pos.getX() + 0.5D, (double)pos.getY(), (double)pos.getZ() + 0.5D, 0.0F, 0.0F);
+					phoenix.setLocationAndAngles(pos.getX() + 0.5D, pos.getY(), pos.getZ() + 0.5D, 0.0F, 0.0F);
 					worldIn.spawnEntityInWorld(phoenix);
 				} else {
 					worldIn.spawnParticle(EnumParticleTypes.EXPLOSION_LARGE, pos.getX(), pos.getY(), pos.getZ(), 0.0D, 0.0D, 0.0D, new int[0]);
